# Tokenized Franchise Performance Management System - Smart Contracts