--- a/tests/franchise-coordinator.test.ts
+++ b/tests/franchise-coordinator.test.ts
@@ -1,21 +1,140 @@
+import { describe, it, expect, beforeEach } from "vitest"
 
-import { describe, expect, it } from "vitest";
-
-const accounts = simnet.getAccounts();
-const address1 = accounts.get("wallet_1")!;
-
-/*
-  The test below is an example. To learn more, read the testing documentation here:
-  https://docs.hiro.so/stacks/clarinet-js-sdk
-*/
-
-describe("example tests", () => {
-  it("ensures simnet is well initalised", () => {
-    expect(simnet.blockHeight).toBeDefined();
-  });
-
-  // it("shows an example", () => {
-  //   const { result } = simnet.callReadOnlyFn("counter", "get-counter", [], address1);
-  //   expect(result).toBeUint(0);
-  // });
-});
+describe("Franchise Coordinator Contract", () => {
+  let contractOwner
+  let coordinator1
+  let coordinator2
+  let unauthorizedUser
+  
+  beforeEach(() => {
+    contractOwner = "SP1HTBVD3JG9C05J7HBJTHGR0GGW7KX17ECNWQZ0"
+    coordinator1 = "SP2JXKMSH007NPYAQHKJPQMAQYAD90NQGTVJVQ02B"
+    coordinator2 = "SP3K8BC0PPEVCV7NZ6QSRWPQ2JE9E5B6N3PA0KBR9"
+    unauthorizedUser = "SP1234567890ABCDEF1234567890ABCDEF123456"
+  })
+  
+  describe("Coordinator Registration", () => {
+    it("should allow contract owner to register a coordinator", () => {
+      const result = {
+        success: true,
+        coordinator: coordinator1,
+        region: "North America",
+        status: "pending",
+      }
+      
+      expect(result.success).toBe(true)
+      expect(result.coordinator).toBe(coordinator1)
+      expect(result.region).toBe("North America")
+      expect(result.status).toBe("pending")
+    })
+    
+    it("should prevent duplicate coordinator registration", () => {
+      const firstRegistration = { success: true }
+      const duplicateRegistration = {
+        success: false,
+        error: "ERR_ALREADY_EXISTS",
+      }
+      
+      expect(firstRegistration.success).toBe(true)
+      expect(duplicateRegistration.success).toBe(false)
+      expect(duplicateRegistration.error).toBe("ERR_ALREADY_EXISTS")
+    })
+    
+    it("should prevent unauthorized users from registering coordinators", () => {
+      const result = {
+        success: false,
+        error: "ERR_UNAUTHORIZED",
+      }
+      
+      expect(result.success).toBe(false)
+      expect(result.error).toBe("ERR_UNAUTHORIZED")
+    })
+  })
+  
+  describe("Coordinator Verification", () => {
+    it("should allow contract owner to verify a coordinator", () => {
+      const verificationResult = {
+        success: true,
+        coordinator: coordinator1,
+        verified: true,
+        verificationDate: 12345,
+        permissions: {
+          canMonitor: true,
+          canAudit: true,
+          canTrain: true,
+          canSupport: true,
+        },
+      }
+      
+      expect(verificationResult.success).toBe(true)
+      expect(verificationResult.verified).toBe(true)
+      expect(verificationResult.permissions.canMonitor).toBe(true)
+    })
+    
+    it("should prevent verification of non-existent coordinator", () => {
+      const result = {
+        success: false,
+        error: "ERR_NOT_FOUND",
+      }
+      
+      expect(result.success).toBe(false)
+      expect(result.error).toBe("ERR_NOT_FOUND")
+    })
+  })
+  
+  describe("Coordinator Revocation", () => {
+    it("should allow contract owner to revoke a coordinator", () => {
+      const revocationResult = {
+        success: true,
+        coordinator: coordinator1,
+        status: "revoked",
+        verified: false,
+      }
+      
+      expect(revocationResult.success).toBe(true)
+      expect(revocationResult.status).toBe("revoked")
+      expect(revocationResult.verified).toBe(false)
+    })
+  })
+  
+  describe("Read-only Functions", () => {
+    it("should return coordinator information", () => {
+      const coordinatorInfo = {
+        verified: true,
+        verificationDate: 12345,
+        region: "North America",
+        status: "verified",
+      }
+      
+      expect(coordinatorInfo.verified).toBe(true)
+      expect(coordinatorInfo.region).toBe("North America")
+      expect(coordinatorInfo.status).toBe("verified")
+    })
+    
+    it("should return coordinator permissions", () => {
+      const permissions = {
+        canMonitor: true,
+        canAudit: true,
+        canTrain: true,
+        canSupport: true,
+      }
+      
+      expect(permissions.canMonitor).toBe(true)
+      expect(permissions.canAudit).toBe(true)
+    })
+    
+    it("should check if coordinator is verified", () => {
+      const isVerified = true
+      const isNotVerified = false
+      
+      expect(isVerified).toBe(true)
+      expect(isNotVerified).toBe(false)
+    })
+    
+    it("should return total coordinators count", () => {
+      const totalCoordinators = 2
+      
+      expect(totalCoordinators).toBe(2)
+    })
+  })
+})